Add files via upload
Task Board....
Develop a task board; The Taskbar allows you to add simple tasks to the clipboard, each
They contain the job description and target time for the task. The notes containing the task contents will appear on a clipboard
The tree in order of income. A task can also be deleted. The user will fill in the task details in the form and after
Clicking the Add button will add the task to the rest of the taskbar tasks and the form will return to clean.
The tasks are saved in the Browser's Local Storage and are actually saved in JSON format if closed.
The browser and return to the page, tasks that are not completed will load to the screen.
HTML5 page design
 Design by CSS3
 Using Bootstrap
 Writing logic in JavaScript:
 Using Storage Local
DOM Manipulations
 Use of HTML events.
 Using arrays and objects
 Using JSON